modules/textToSpeech/pyttsxModule/requirements.txt
pyzmq>=15.2.0
